backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pandas>=2.3.1",
    "psycopg2-binary>=2.9.10",
    "python-jose[cryptography]>=3.5.0",
    "requests>=2.32.4",
    "sqlmodel>=0.0.24",
    "python-dotenv>=1.0.0",
    "langchain-community>=0.3.27",
    "langchain-core>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain>=0.3.0",
    "supabase>=2.18.1",
]
